Code/requirements.txt
absl-py==1.0.0
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
audioread==2.1.9
Bottleneck==1.3.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.0
dataclasses==0.6
decorator==5.1.1
filelock==3.6.0
Flask==2.0.3
Flask-Cors==3.0.10
flatbuffers==2.0
gast==0.2.2
google-api-core==1.31.5
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-core==1.7.2
google-cloud-translate==2.0.1
google-pasta==0.2.0
googleapis-common-protos==1.55.0
grpcio==1.44.0
h5py==3.6.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libclang==13.0.0
librosa==0.9.1
llvmlite==0.38.0
Markdown==3.3.4
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
nltk==3.7
numba==0.55.1
numexpr==2.8.1
numpy==1.21.5
oauthlib==3.2.0
opt-einsum==3.3.0
ortools==9.3.10497
packaging==21.3
pandas==1.3.4
Pillow==9.0.1
pip==21.2.2
pooch==1.6.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-speech-features==0.6
pytz==2021.3
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.0
sentencepiece==0.1.96
setuptools==58.0.4
six==1.16.0
SoundFile==0.10.3.post1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-estimator==2.0.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tokenizers==0.11.6
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.1
transformers==4.17.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wheel==0.37.1
wrapt==1.13.3
zipp==3.7.0
